h265 code check Signed-off-by: s30029175 <shiqiwei4@huawei.com>

+++ services/engine/gstreamer/plugins/codec/common/vdec/gst_vdec_h265.cpp
@@ -104,7 +104,7 @@ static gboolean gst_vdec_h265_parser_nalu(GstVdecH265 *self, const GstMapInfo &s
         p_dts_data[3] = 1;
 
         guint byte_count = 0;
-        guint hvcc_nal_size = self->hvcc_nal_len;
+        guint hvcc_nal_size = static_cast<guint>(self->hvcc_nal_len);
         for (guint i = 0; i < hvcc_nal_size; ++i) {
             // two hexadecimal code in 1 byte(8 bit)
             byte_count = (byte_count << 8) + p_src_data[i];
